# modern-api-book/converting_openapi_to_reactive_api_application/src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: Sample Ecommerce App
  description: >
    'This is a ***sample ecommerce app API***.  You can find out more about Swagger at [swagger.io](http://swagger.io).
    Description supports markdown markup. For example, you can use the `inline code` using back ticks.'
  termsOfService: https://github.com/PacktPublishing/Modern-API-Development-with-Spring-and-Spring-Boot/blob/master/LICENSE
  contact:
    email: support@packtpub.com
  license:
    name: MIT
    url: https://github.com/PacktPublishing/Modern-API-Development-with-Spring-and-Spring-Boot/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/token:
    post:
      tags:
        - user
      summary: Signin the customer (user)
      description: Generates the JWT and refresh token
      operationId: signIn
      requestBody:
        content:
          application/xml:
            schema:
              $ref: "#/components/schemas/SignInReq"
          application/json:
            schema:
              $ref: "#/components/schemas/SignInReq"
      responses:
        200:
          description: Returns the access and refresh token
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/SignedInUser"
            application/json:
              schema:
                $ref: "#/components/schemas/SignedInUser"
    delete:
      tags:
        - user
      summary: Signouts the customer (user)
      description: Signouts the customer (user)
      operationId: signOut
      requestBody:
        content:
          application/xml:
            schema:
              $ref: "#/components/schemas/RefreshToken"
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshToken"
      responses:
        202:
          description: Accetps the request for logout
          content: {}
  /api/v1/auth/token/refresh:
    post:
      tags:
        - user
      summary: Provides new JWT based on valid refresh token.
      description: Provides new JWT based on valid refresh token.
      operationId: getAccessToken
      requestBody:
        content:
          application/xml:
            schema:
              $ref: "#/components/schemas/RefreshToken"
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshToken"

      responses:
        200:
          description: For successful operation
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/SignedInUser"
            application/json:
              schema:
                $ref: "#/components/schemas/SignedInUser"
        400:
          description: refresh token not exist
          content: {}
...
  /api/v1/products:
    get:
      tags:
        - product
      summary: Returns all the matched products
      description: Returns the products that matches the given query criteria
      operationId: queryProducts
      parameters:
        - name: tag
          in: query
          description: Product tag
          required: false
          schema:
            type: string
        - name: name
          in: query
          description: Product name
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Query page number
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          description: Query page size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        200:
          description: For successful fetch.
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/v1/products/{id}:
    get:
      tags:
        - product
      summary: Returns a product
      description: Returns the product that matches the given product ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          description: Product Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
components:
  schemas:
...
    BigDecimal:
      type: BigDecimal
...
    Product:
      description: Product information
      type: object
      properties:
        id:
          description: Product identifier
          type: string
        name:
          description: Product Name
          type: string
        description:
          description: Prodcut's description
          type: string
        imageUrl:
          description: Product image's URL
          type: string
        price:
          description: Product price
          $ref: '#/components/schemas/BigDecimal'
        count:
          description: Product count
          type: integer
          format: int32
        tag:
          description: Tags associated with the product
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    Tag:
      description: Tag information
      type: object
      properties:
        id:
          description: Tag identifier
          type: string
        name:
          description: Tag Name
          type: string
    SignInReq:
      description:  Request body for sign-in
      type: object
      properties:
        username:
          description: username of the User
          type: string
        password:
          description: password of the User
          type: string
    SignedInUser:
      description: Signed-in User Information
      type: object
      properties:
        refreshToken:
          description: Contains the Refresh Token
          type: string
        accessToken:
          description: Contains the JWT Access Token
          type: string
        username:
          description: User Name
          type: string
        userId:
          description: User Identifier
          type: string
    RefreshToken:
      description: Contains the Refresh Token
      type: object
      properties:
        refreshToken:
          description: Refresh Token
          type: string